Set dtype explicitly on DataFrame creation
When the DataFrame for a song was initialized, it would default its
dtype to 'object', creating typing issues when the data was successively
added and manipulated. Song.get_plays() would raise a TypeError when
trying to interpolate the data points, for example.

+++ kstreams/classes.py
@@ -269,7 +269,7 @@ class SongDB(object):
                                        'agency': songinfo['agency']}
 
         dbpath = os.path.join(self.path, '{}.pkl'.format(songinfo['id']))
-        pd.DataFrame(columns=['plays', 'listeners']).to_pickle(dbpath)
+        pd.DataFrame(columns=['plays', 'listeners'], dtype=int).to_pickle(dbpath)
         logging.debug('Added to database (%s by %s)',
                       songinfo['title'], songinfo['artist'])
